[Update] Intercept now supports "auto" return
Returning the string "auto" will cause the indent to be the same as the
next node included. If no node is included until the end, or until go
back one level, the child is included in the current parent (it is, as
sibling of other items in the current level). This behavior is better
described as putting the node together in the sibling level.