[Fix] lineNum without emptyLines option
Now the line number is realiable including when emptyLines is false.

An unpredictable case is when emptyLines is activated, in this case the
Node lineNum will be of the last uncached node. It is not a problem,
once when emptyLines is true the Node instances must be ignored, only
caring for Leaf instances.

**Quick fixes**
In the future, it is easy to make a quick fix for the case related
above. It must only decrease the cache.length from the Node.lineNum and
the expected line number will shows up. Not a thing to care for now.